fix: ensure there is no overlap between implementations for node and browser
